Adding works (edit is broken)

+++ INTEX_II_Group_4-3/Models/TopProductRecommendation.cs
@@ -8,12 +8,12 @@ namespace INTEX_II_Group_4_3.Models
 
         [Key]
         [ForeignKey("ProductRec")]
-        public double product_ID { get; set; }
+        public int product_ID { get; set; }
         //ProductRec is used for referencing in the views
         public Product ProductRec { get; set; }
 
         //[ForeignKey("Product_1")]
-        public double ratings_count { get; set; }
+        public int ratings_count { get; set; }
         //public Product Product_1 { get; set; }
 
         //[ForeignKey("Product_2")]
